+++ vm_core.h
@@ -727,69 +727,14 @@ typedef struct rb_vm_struct {
             // join at exit
             rb_nativethread_cond_t terminate_cond;
             bool terminate_waiting;
-
-#ifndef RUBY_THREAD_PTHREAD_H
-            // win32
-            bool barrier_waiting;
-            unsigned int barrier_cnt;
-            rb_nativethread_cond_t barrier_complete_cond;
-            rb_nativethread_cond_t barrier_release_cond;
-#endif
         } sync;
 
         /* VM-wide locks for the Ractor transfer/inheritance machinery.  All of them
          * are leaf locks: no safepoint inside a critical section. */
         rb_nativethread_lock_t generic_fields_lock;   /* the shared generic-fields table in variable.c */
 
-#ifdef RUBY_THREAD_PTHREAD_H
-        // ractor scheduling
-        struct {
-            rb_nativethread_lock_t lock;
-            struct rb_ractor_struct *lock_owner;
-            bool locked;
-
-            rb_nativethread_cond_t cond; // GRQ
-            rb_atomic_t snt_cnt;  // count of shared NTs; lock-free (see native_thread_dedicated_inc)
-            unsigned int dnt_cnt; // count of dedicated NTs; logging only (USE_RUBY_DEBUG_LOG), not atomic
-
-
-            unsigned int max_cpu;
-            struct ccan_list_head grq; // // Global Ready Queue
-            rb_atomic_t winding_cnt; // native threads between a coroutine epilogue and its reclaim; ruby_vm_destruct waits for 0
-            unsigned int grq_cnt;
-
-            // What the barrier walk visits: threads running on dedicated
-            // nts, and the shared nts (whose running_th fields hold the rest).
-            struct {
-                rb_nativethread_lock_t lock;
-                struct ccan_list_head running_dnts;
-                struct ccan_list_head snts;
-            } ntlist;
-
-            // scheds whose readyq holds waiters: the timer ticks their
-            // running thread (timeslice_scan) and prunes drained entries.
-            struct {
-                rb_nativethread_lock_t lock;
-                struct ccan_list_head scheds;
-            } timeslice;
-
-            // true if timeslice timer is not enable
-            bool timeslice_wait_inf;
-
-            // barrier
-            rb_nativethread_cond_t barrier_complete_cond;
-            rb_nativethread_cond_t barrier_release_cond;
-            // bool; nonzero while a stop-the-world section is active.  Set
-            // before the barrier walks the running records; a record moved
-            // after the walk sees it (thread_sched_setup_running_threads).
-            rb_atomic_t barrier_is_waiting;
-            unsigned int barrier_joined_cnt; // threads joined so far; under sched.lock
-            unsigned int barrier_running_cnt; // runners counted by the barrier's walk; under sched.lock
-            unsigned int barrier_serial;
-            struct rb_ractor_struct *barrier_ractor;
-            unsigned int barrier_lock_rec;
-        } sched;
-#endif
+        // ractor scheduling; see thread_sched.h
+        struct rb_ractor_sched sched;
     } ractor;
 
 #ifdef USE_SIGALTSTACK
