+++ vm_sync.c
@@ -65,21 +65,13 @@ rb_vm_locked_p(void)
 static bool
 vm_need_barrier_waiting(const rb_vm_t *vm)
 {
-#ifdef RUBY_THREAD_PTHREAD_H
     return vm->ractor.sched.barrier_is_waiting;
-#else
-    return vm->ractor.sync.barrier_waiting;
-#endif
 }
 
 static bool
 vm_need_barrier(bool no_barrier, const rb_ractor_t *cr, const rb_vm_t *vm)
 {
-#ifdef RUBY_THREAD_PTHREAD_H
     return !no_barrier && cr->threads.sched.running != NULL && vm_need_barrier_waiting(vm); // ractor has running threads.
-#else
-    return !no_barrier && vm_need_barrier_waiting(vm);
-#endif
 }
 
 static void
@@ -142,13 +134,11 @@ vm_lock_leave(rb_vm_t *vm, bool no_barrier, unsigned int *lev APPEND_LOCATION_AR
     VM_ASSERT(vm->ractor.sync.lock_rec == *lev);
     VM_ASSERT(cr == GET_RACTOR());
 
-#ifdef RUBY_THREAD_PTHREAD_H
     if (vm->ractor.sched.barrier_ractor == cr &&
         vm->ractor.sched.barrier_lock_rec == vm->ractor.sync.lock_rec) {
         VM_ASSERT(!no_barrier);
         rb_ractor_sched_barrier_end(vm, cr);
     }
-#endif
 
     if (RUBY_DTRACE_GVL_RELEASE_ENABLED()) {
         RUBY_DTRACE_GVL_RELEASE();
@@ -258,11 +248,7 @@ rb_vm_cond_timedwait(rb_vm_t *vm, rb_nativethread_cond_t *cond, unsigned long ms
 static bool
 vm_barrier_acquired_p(const rb_vm_t *vm, const rb_ractor_t *cr)
 {
-#ifdef RUBY_THREAD_PTHREAD_H
     return vm->ractor.sched.barrier_ractor == cr;
-#else
-    return false;
-#endif
 }
 
 void
